fix batch api
